+++ package.json
@@ -26,7 +26,8 @@
     "next": "11.0.1",
     "query-string": "^7.0.1",
     "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react-dom": "17.0.2",
+    "swr": "^0.5.6"
   },
   "devDependencies": {
     "eslint": "7.30.0",
