+++ styles/Home.module.css
@@ -20,7 +20,7 @@
 
 .footer {
   width: 100%;
-  height: 100px;
+  padding: 24px;
   border-top: 1px solid #eaeaea;
   display: flex;
   justify-content: center;
